pyproject.toml
[project]
name = "azure-distributed-trace-azure-monitor"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "opentelemetry-api (>=1.35.0,<2.0.0)",
    "opentelemetry-sdk (>=1.35.0,<2.0.0)",
    "azure-monitor-opentelemetry-exporter (>=1.0.0b40,<2.0.0)",
    "azure-core (>=1.35.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.30.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
